Add not-test demonstrating schedule problems
